+++ src/wrapper/oracle.cpp
