+++ src/wrapper/rpc/wrapper_rpc.cpp
@@ -1,7 +1,7 @@
-#include "oracle.h"
 #include "quartz/tasograph/substitution.h"
 #include "quartz/tasograph/tasograph.h"
 #include "rpc/server.h"
+#include "wrapper/oracle.h"
 
 #include <thread>
 using namespace quartz;
@@ -16,7 +16,7 @@ int main(int argc, char **argv) {
   float timeout = atof(argv[5]);
   int nqubits = atoi(argv[6]);
   rpc::server srv(port);
-  auto supercontext = get_context_(argv[2], nqubits, argv[3]);
+  auto supercontext = get_context_(argv[2], 1, argv[3]);
   std::cout << "Server started on port " << port << std::endl;
   srv.bind("optimize", [supercontext, timeout, argv](std::string my_circ) {
     return optimize_(my_circ, argv[4], timeout, supercontext);
